bot/requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
future==0.16.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gspread==3.0.1
httplib2==0.11.3
idna==2.7
oauth2client==4.1.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-telegram-bot==10.1.0
requests==2.19.1
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
